JSON
newe

+++ app/routes/metodosJSON.js
@@ -1,5 +1,6 @@
 var fs = require("fs");
-const aes = require('./aes');
+const aes = require("./aes");
+var path = require("path");
 function crearJSON(idpac,idmed,idchat){
     let arch = "p"+idpac+"m"+idmed+"c"+idchat+".json";
     var obj = {
@@ -10,12 +11,17 @@ function crearJSON(idpac,idmed,idchat){
     fs.writeFile(arch,json,"utf-8",callback);
 }
 function escribirJSON(idpac,idmed,idchat,tipusr, msg){
-    var obj = {
-        table:[]
-    };
     let arch = "p"+idpac+"m"+idmed+"c"+idchat+".json";
-    obj.table.push({tip_user:aes.cifrar(tipusr),msg:aes.cifrar(msg)});
-    fs.writeFile(arch,json,"utf-8",callback);
+    if(!path.exists(arch)){
+        crearJSON(idpac,idmed,idchat);
+    }else{
+        var obj = {
+        table:[]
+        };
+        obj.table.push({tip_user:aes.cifrar(tipusr),msg:aes.cifrar(msg)});
+        fs.writeFile(arch,json,"utf-8",callback);
+    }
+    
 }
 exports.crearJSON = crearJSON;
 exports.escribirJSON = escribirJSON;
